openapi directory renamed to spec (for codeclimate)
