+++ common-libs/internal-api-clients/src/services/BloomVaultApiService.ts
@@ -1,6 +1,6 @@
 import { profile } from '@affinidi/common'
 
-import bloomVaultSpec from '../openapi/_bloomVault'
+import bloomVaultSpec from '../spec/_bloomVault'
 import GenericApiService, { GenericConstructorOptions } from './GenericApiService'
 import { ParseSpec } from '../types/openapiParser'
 import { BuildApiType } from '../types/typeBuilder'
