+++ common-libs/internal-api-clients/src/spec/_affinidiVault.ts
