+++ common-libs/internal-api-clients/src/spec/_revocation.ts
